fix: hook up game url with real game id
