+++ index.html
@@ -3,10 +3,10 @@
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <link rel="preload" as="fetch" href="websockets-doom.wasm">
-    <link rel="preload" as="fetch" href="freedoom2.wad">
-    <link rel="preload" as="fetch" href="Cardano.wad">
-    <link rel="preload" as="fetch" href="default.cfg">
+    <link rel="preload" as="fetch" href="websockets-doom.wasm" />
+    <link rel="preload" as="fetch" href="freedoom2.wad" />
+    <link rel="preload" as="fetch" href="Cardano.wad" />
+    <link rel="preload" as="fetch" href="default.cfg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Hydra Doom Multiplayer</title>
   </head>
