+++ src/components/GameView/GameView.tsx
@@ -1,5 +1,3 @@
-import { MdContentCopy } from "react-icons/md";
-import { ClipboardAPI, useClipboard } from "use-clipboard-copy";
 import Card from "../Card";
 import DoomCanvas from "../DoomCanvas";
 import GlobalTPS from "../GlobalTPS";
@@ -9,24 +7,8 @@ import MusicPlayer from "../MusicPlayer";
 import RestartButton from "../RestartButton";
 import StatsCard from "../StatsCard";
 import TopLinks from "../TopLinks";
-import { useCallback } from "react";
-import { FaRegCircleCheck } from "react-icons/fa6";
-import { useAppContext } from "../../context/useAppContext";
-import { EGameType } from "../../types";
 
 const GameView = () => {
-  const { gameData } = useAppContext();
-  const urlClipboard = useClipboard({ copiedTimeout: 1500 });
-
-  const urlClipboardCopy = useCallback(
-    (clipboard: ClipboardAPI, value: string) => {
-      clipboard.copy(value);
-    },
-    [],
-  );
-
-  const gameUrl = `${window.location.origin}/join/1efabc`;
-
   return (
     <Layout>
       <TopLinks />
@@ -50,29 +32,7 @@ const GameView = () => {
           <HydraHeadLiveTxs />
         </div>
         <div className="flex flex-col gap-6">
-          <Card className="h-[40rem]">
-            <DoomCanvas />
-          </Card>
-          {gameData.type === EGameType.HOST && (
-            <Card className="px-4 py-2 text-center text-xl text-white flex items-center gap-2 justify-center">
-              Share this URL with friends{" "}
-              <a
-                className="text-yellow-400 underline"
-                href={gameUrl}
-                target="_blank"
-              >
-                {gameUrl}
-              </a>
-              {urlClipboard.copied ? (
-                <FaRegCircleCheck className="text-green-600" />
-              ) : (
-                <MdContentCopy
-                  role="button"
-                  onClick={() => urlClipboardCopy(urlClipboard, gameUrl)}
-                />
-              )}
-            </Card>
-          )}
+          <DoomCanvas />
         </div>
         <div className="w-80 flex flex-col gap-4">
           <GlobalTPS size="sm" titleAlign="left" />
